refactor: move model class utils to `graph` directory
